ml/scripts/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
tqdm==4.66.5
sentence-transformers==3.0.1
torch==2.4.1
numpy==1.26.4
